refactor(Controller): modulize controller
